requirements.txt
absl-py==0.9.0
amqp==2.5.2
asgiref==3.2.3
astor==0.8.1
attrs==19.3.0
autopep8==1.4.4
billiard==3.6.2.0
bleach==1.5.0
blis==0.2.4
cachetools==4.0.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
Django==2.1
django-cors-headers==3.2.1
django-loadjson==0.1.2
djangorestframework==3.11.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
html5lib==0.9999999
idna==2.8
importlib-metadata==1.5.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
kombu==4.6.7
Markdown==3.2.1
MarkupSafe==1.1.1
mathparse==0.1.2
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==5.2.2
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.24.2
pandocfilters==1.4.2
pickleshare==0.7.5
Pillow==7.0.0
Pint==0.9
plac==0.9.6
portpicker==1.2.0
preshed==2.0.1
prompt-toolkit==1.0.18
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.5.2
pypiwin32==219
pyrsistent==0.15.7
python-dateutil==2.7.5
pytz==2019.3
pywin32==227
PyYAML==5.1.2
pyzmq==19.0.0
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
soupsieve==2.0
sqlparse==0.3.0
srsly==1.0.1
tensorboard==2.1.0
tensorflow==1.5.0
tensorflow-estimator==2.1.0
tensorflow-serving-api-python3==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
testpath==0.4.4
tflearn==0.3.2
thinc==7.0.8
tornado==4.5.3
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.24.3
vine==1.3.0
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
wasabi==0.6.0
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==1.0.0
whitenoise==5.0.1
win-unicode-console==0.5
wrapt==1.12.0
zipp==1.2.0
